dom/events/test/test_bug742376.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=402089
-->
<head>
  <title>Test for Bug 742376</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>

<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=742376">Mozilla Bug 742376</a>
<script class="testbody" type="text/javascript">

/** Test for Bug 742376 **/

function dumbListener(event) {}
function dumbListener2(event) {}
function dumbListener3(event) {}

window.addEventListener("deviceorientation", dumbListener, false);
window.addEventListener("random_event_name", function() {}, false);
window.addEventListener("deviceorientation", dumbListener2, false);

window.setTimeout(function() {

  window.removeEventListener("deviceorientation", dumbListener, false);
  window.setTimeout(function() {

    window.removeEventListener("deviceorientation", dumbListener2, false);
    window.setTimeout(function() {
      testEventHandler();
    }, 0);
  }, 0);
}, 0);

function testEventHandler() {
  window.ondeviceorientation = function() {}
  window.setTimeout(function() {
    window.ondeviceorientation = null;
    window.setTimeout(function() {
      SimpleTest.finish();
    }, 0);
  }, 0)
}

SimpleTest.waitForExplicitFinish();

</script>
</pre>
</body>
</html>
